:bug: Fix timeout types in bucket creation

+++ {{cookiecutter.project_slug}}/backend/app/app/db/database.py
@@ -41,8 +41,8 @@ def get_bucket(
     bucket_name: str,
     host="couchbase",
     port="8091",
-    timeout: int = COUCHBASE_OPERATION_TIMEOUT_SECS,
-    n1ql_timeout: int = COUCHBASE_N1QL_TIMEOUT_SECS,
+    timeout: float = COUCHBASE_OPERATION_TIMEOUT_SECS,
+    n1ql_timeout: float = COUCHBASE_N1QL_TIMEOUT_SECS,
 ):
     cluster = get_cluster(username, password, host=host, port=port)
     bucket: Bucket = cluster.open_bucket(bucket_name, lockmode=LOCKMODE_WAIT)
